requirements.txt
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
psycopg2==2.9.1
pycparser==2.20
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-magic==0.4.24
requests==2.25.1
urllib3==1.26.6
